Bugfix: PGPs display as float instead of int
PGPs were calculated as int, so sessions less than one hour were
calculated as zero. This fix returns a timedelta as a float instead of a
`divmod` calculation.

Schemas updated to display event total as a float during serialization.

+++ app/schemas.py
@@ -100,7 +100,7 @@ class CourseAttendingSchema(Schema):
     type = fields.Nested(CourseTypeSchema)
     presenters = fields.Nested("CoursePresenterSchema", many=True)
     links = fields.Nested("DisplayCourseLinkSchema", many=True)
-    total = fields.Int()
+    total = fields.Float()
     available = fields.Int()
